Moved classes out and created protocols
